issue-1038: splitting DsQuality into (DsComment, DsQuality, DsPk)
